[parser] Move the parser from `lib` to `sh`
The files in `lib` are concatenated and runned on each and every page.
However, the parser is a development-only tool,
so it is not needed so often.
