+++ lib/load.js
@@ -30,7 +30,6 @@
 		'lib/core_dvig.js',
 		'lib/dvig_fn.js',
 		'lib/dvig_rz.js',
-		'lib/dvig_lz.js',
 		'lib/urljson.js',
 		'lib/style.js',
 		'lib/menu.js',
