+++ sh/dvig_lz.js
